requirements.txt
faker
typer
pyYAML
